Exception added, minor code cleanups
- Exception class added to base/common.h. TODO: report critical errors via
  exceptions
- (missing) static and const qualifiers added to some member functions,
  Hash*::cached now mutable, so getters are now const, other minor code
  cleanups (redundant includes, more explicit casts ...)
- flags for profiling builds added to CMakeFile.txt
- Audio closer to workable state
- sliding works better

+++ src/client/MD2.h
@@ -46,9 +46,9 @@ namespace client
 
       struct Anim
       {
-        int firstFrame;
-        int lastFrame;
-        int fps;
+        int  firstFrame;
+        int  lastFrame;
+        float fps;
       };
 
       struct AnimState
@@ -56,7 +56,7 @@ namespace client
         int   startFrame;
         int   endFrame;
 
-        int   fps;
+        float fps;
         float frameTime;
         float currTime;
         float oldTime;
